+++ vehicles/seat_1p/__init__.py
@@ -1,1 +1,0 @@
-"""SEAT Leon 1P vehicle profile assets."""